requirements.txt
streamlit
google-generativeai
faiss-cpu
sentence-transformers 